scripts_bucket/lib/requirements.txt
google-auth
google-cloud-retail
google-cloud-bigquery
google-cloud-storage
google-cloud-core
google-api-python-client
grpcio
protobuf